Add debug configuration & *.launch
https://e2e.ti.com/support/tools/code-composer-studio-group/ccs/f/code-composer-studio-forum/737306/

Basically, enabling that option is telling CCS that you want the CCS ccs-ccstudio-istargetmanual-flag-meaning-in-the-ccsproject-file
project manager to maintain the target configuration file.
For the project manager to do so, it needs to know the connection type.
Hence why you can choose the connection type.

If you uncheck the option, you are telling CCS that you will maintain
the target configuration file yourself manually. Hence the connection
type is not needed by the project manager since you will maintain the
target configuration yourself. Hence the option to choose in the project
is disabled.